# api/openapi.yaml
openapi: 3.0.3
info:
  title: Digital Asset Wallet Platform API
  version: "1.0"
  description: >
    Internal API for the digital asset wallet platform. Every mutating route requires
    bearer authentication and an Idempotency-Key header (AD-14). Errors use RFC 9457
    application/problem+json.
servers:
  - url: http://localhost:8080/v1
    description: Local dev (docker compose / make run)
  - url: /v1
security:
  - bearerAuth: []
paths:
...
  /transfers:
    post:
      operationId: createTransfer
      summary: Move balance from one customer to another, ledger-only, no chain interaction
      description: >
        A single balanced journal entry (debit source, credit destination) commits
        atomically within one (chain, asset) pair — the same granularity as the
        provisioned per-customer accounts (FR4, AD-3, AD-4). No on-chain transaction is
        ever created.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferRequest"
      responses:
        "201":
          description: Transfer completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transfer"
        "400":
          description: >
            Bad request — missing Idempotency-Key, malformed body, invalid chain/asset,
            non-positive amount, or source equals destination
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Missing or invalid bearer token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Source or destination customer has no account for the requested (chain, asset)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "409":
          description: >
            Idempotency-Key reused with a different request body, or a concurrent
            request already used this key to write a journal entry
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "422":
          description: Source account balance is less than the requested amount
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          description: Unexpected server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
...
  /customers/{id}/balances:
    get:
      operationId: getCustomerBalances
      summary: Query a customer's current balance for each supported asset and chain
      description: >
        Each balance is derived from the ledger's postings, never a stored value (AD-3).
        Amounts are integer base units — wei for ETH, 6-decimal units for USDC — encoded
        as JSON strings, never numbers, to avoid float64 precision loss.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Customer balances
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BalancesResponse"
        "401":
          description: Missing or invalid bearer token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: No customer with the given id exists
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          description: Unexpected server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
...
  /withdrawals/fee-estimate:
    get:
      operationId: getWithdrawalFeeEstimate
      summary: Query a withdrawal's estimated fee before submitting one
      description: >
        Combines the L2 execution fee and the amortized L1 data-posting fee for the given
        chain (Arbitrum via NodeInterface.gasEstimateComponents(), Base via the
        GasPriceOracle predeploy) — a naive single-number gas estimate would systematically
        undercharge by omitting the L1 component entirely, so both components are always
        returned explicitly alongside their sum, never collapsed into one undifferentiated
        number. No withdrawal resource exists until Story 3.2 — this is a pure,
        unpersisted, read-only computation; nothing is written.
      parameters:
        - name: chain
          in: query
          required: true
          schema:
            type: string
            enum: [base, arbitrum]
        - name: asset
          in: query
          required: true
          schema:
            type: string
            enum: [eth, usdc]
        - name: amount
          in: query
          required: true
          description: >
            Integer base units (wei for ETH, 6-decimal units for USDC), encoded as a
            string — matching Balance's convention. Must be a positive integer no larger
            than a uint256. Validated but not itself a factor in the returned fee: gas
            cost for a plain transfer or token transfer call does not vary with the
            transferred amount, so the same fee is returned for any valid amount of a
            given chain/asset pair.
          schema:
            type: string
      responses:
        "200":
          description: The withdrawal's estimated fee, split into its L2 and L1 components
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeeEstimate"
        "400":
          description: >
            Invalid chain or asset enum value, or amount is missing, non-positive, or not
            a base-10 integer
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Missing or invalid bearer token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          description: >
            Unexpected server error — e.g. the requested chain/asset combination has no
            token_registry entry for USDC (a misconfigured deployment, never a guessed
            contract address)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
  schemas:
...
    Balance:
      type: object
      required:
        - chain
        - asset
        - balance
      properties:
        chain:
          type: string
          enum: [base, arbitrum]
        asset:
          type: string
          enum: [eth, usdc]
        balance:
          type: string
          description: >
            Integer base units (wei for ETH, 6-decimal units for USDC), encoded as a
            string — never a JSON number — to avoid float64 precision loss.
    BalancesResponse:
      type: object
      required:
        - balances
      properties:
        balances:
          type: array
          items:
            $ref: "#/components/schemas/Balance"
    TransferRequest:
      type: object
      required:
        - sourceCustomerId
        - destinationCustomerId
        - chain
        - asset
        - amount
      properties:
        sourceCustomerId:
          type: string
          format: uuid
        destinationCustomerId:
          type: string
          format: uuid
        chain:
          type: string
          enum: [base, arbitrum]
        asset:
          type: string
          enum: [eth, usdc]
        amount:
          type: string
          description: >
            Integer base units (wei for ETH, 6-decimal units for USDC), encoded as a
            string — never a JSON number — matching Balance's convention. Must be a
            positive integer.
    Transfer:
      type: object
      required:
        - id
        - sourceCustomerId
        - destinationCustomerId
        - chain
        - asset
        - amount
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        sourceCustomerId:
          type: string
          format: uuid
        destinationCustomerId:
          type: string
          format: uuid
        chain:
          type: string
          enum: [base, arbitrum]
        asset:
          type: string
          enum: [eth, usdc]
        amount:
          type: string
        createdAt:
          type: string
          format: date-time
...
    FeeEstimate:
      type: object
      required:
        - l2Fee
        - l1Fee
        - totalFee
      properties:
        l2Fee:
          type: string
          description: >
            The L2 execution fee component, in integer base units (wei), encoded as a
            string — never a JSON number — matching Balance's convention.
        l1Fee:
          type: string
          description: >
            The amortized L1 data-posting fee component, in integer base units (wei),
            encoded as a string — never a JSON number.
        totalFee:
          type: string
          description: >
            l2Fee + l1Fee, in integer base units (wei), encoded as a string — never a JSON
            number, and never a naive single-number estimate collapsing the two
            components.
    # RFC 9457 (application/problem+json). Every error response from this service
    # uses this shape, starting this story — see internal/adapter/api/problem.go.
    ProblemDetails:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string